Support correctly weird characters in filenames when downloading
This commit add support for the following:
- Correctly encoded oData parameter sent to the endpoint.
- Support for # and % in filenames by using a newer 365 endpoint.

Summary:

Current implementation was injecting an URL directly quoted to
the endpoint, instead of encoding as an oData value. This was
solved by using the encode_method_value method. This fixes
issues when the file contains weird characters, like slashes
and single quotes.

The getFileServerByUrl endpoint does not support some characters
in filenames, like # and %. A newer endpoint
(getFileByServerRelativePath) is used instead, which allows to
download files with those characters.