Revert ClientObject.get_property() semantics.
In commit 2f486ed4fc775b02224ea38242a9f911cd329a9e a refactoring
of get_property introduced an incompatible change. In that version
a property that was added with set_property() is not retrievable,
as the access to the _properties underlying object uses the
normalized name, instead of the name.

This change adds the previouse behavior (that was right, in my
opinion) in order to not to break backwards compatibility.